+++ packages/amplify-cli/src/extensions/amplify-helpers/trigger-flow.js
@@ -140,7 +140,7 @@ const updateTrigger = async triggerOptions => {
         triggerEventPath,
         skipEdit,
       },
-      functionName
+      functionName,
     );
     if (values && values.length > 0) {
       for (let v = 0; v < values.length; v += 1) {
@@ -217,7 +217,7 @@ const triggerFlow = async (context, resource, category, previousTriggers = {}) =
     return null;
   }
 
-  const pluginPath = context.amplify.getCategoryPlugins(context)[category];
+  const pluginPath = context.amplify.getCategoryPluginInfo(context, category).packageLocation;
 
   // path to trigger directory in category
   const triggerPath = `${pluginPath}/provider-utils/awscloudformation/triggers/`;
@@ -271,7 +271,7 @@ const triggerFlow = async (context, resource, category, previousTriggers = {}) =
         delete triggerObj[Object.keys(tempTriggerObj)[index]];
       }
     },
-    { triggerObj }
+    { triggerObj },
   );
   return triggerObj;
 };
@@ -298,7 +298,8 @@ const getTriggerPermissions = async (context, triggers, category) => {
   let permissions = [];
   const parsedTriggers = JSON.parse(triggers);
   const triggerKeys = Object.keys(parsedTriggers);
-  const pluginPath = context.amplify.getCategoryPlugins(context)[category];
+
+  const pluginPath = context.amplify.getCategoryPluginInfo(context, category).packageLocation;
 
   for (let c = 0; c < triggerKeys.length; c += 1) {
     const index = triggerKeys[c];
@@ -328,7 +329,7 @@ const learnMoreLoop = async (key, map, metaData, question) => {
     if (metaData.URL) {
       prefix = `\nAdditional information about the ${key} available for ${map} can be found here: ${chalkpipe(
         null,
-        chalk.blue.underline
+        chalk.blue.underline,
       )(metaData.URL)}\n`;
       prefix = prefix.concat('\n');
     } else {
@@ -336,7 +337,7 @@ const learnMoreLoop = async (key, map, metaData, question) => {
       Object.values(metaData).forEach(m => {
         prefix = prefix.concat('\n');
         prefix = prefix.concat(
-          `${chalkpipe(null, chalk.green)('\nName:')} ${m.name}${chalkpipe(null, chalk.green)('\nDescription:')} ${m.description}\n`
+          `${chalkpipe(null, chalk.green)('\nName:')} ${m.name}${chalkpipe(null, chalk.green)('\nDescription:')} ${m.description}\n`,
         );
         prefix = prefix.concat('\n');
       });
@@ -379,8 +380,8 @@ const copyFunctions = async (key, value, category, context, targetPath) => {
       fs.mkdirSync(targetPath);
     }
     const dirContents = fs.readdirSync(targetPath);
-    const pluginPath = context.amplify.getCategoryPlugins(context)[category];
-    const functionPath = context.amplify.getCategoryPlugins(context).function;
+    const pluginPath = context.amplify.getCategoryPluginInfo(context, category).packageLocation;
+    const functionPath = context.amplify.getCategoryPluginInfo(context, 'function').packageLocation;
 
     if (!dirContents.includes(`${value}.js`)) {
       let source = '';
@@ -398,7 +399,7 @@ const copyFunctions = async (key, value, category, context, targetPath) => {
 };
 
 const cleanFunctions = async (key, values, category, context, targetPath) => {
-  const pluginPath = context.amplify.getCategoryPlugins(context)[category];
+  const pluginPath = context.amplify.getCategoryPluginInfo(context, category).packageLocation;
   try {
     const meta = context.amplify.getTriggerMetadata(`${pluginPath}/provider-utils/awscloudformation/triggers/${key}`, key);
     const dirContents = fs.readdirSync(targetPath);
@@ -431,7 +432,7 @@ const cleanFunctions = async (key, values, category, context, targetPath) => {
 };
 
 const getTriggerEnvVariables = (context, trigger, category) => {
-  const pluginPath = context.amplify.getCategoryPlugins(context)[category];
+  const pluginPath = context.amplify.getCategoryPluginInfo(context, category).packageLocation;
   let env = [];
   const meta = context.amplify.getTriggerMetadata(`${pluginPath}/provider-utils/awscloudformation/triggers/${trigger.key}`, trigger.key);
   if (trigger.modules) {