+++ packages/amplify-category-interactions/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-interactions@2.3.1...amplify-category-interactions@2.3.2) (2020-03-10)
+
+**Note:** Version bump only for package amplify-category-interactions
+
+
+
+
+
+## [2.3.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-interactions@2.1.11...amplify-category-interactions@2.3.1) (2020-03-07)
+
+**Note:** Version bump only for package amplify-category-interactions
+
+
+
+
+
+## [2.2.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-interactions@2.1.13-beta.0...amplify-category-interactions@2.2.1) (2020-03-05)
+
+**Note:** Version bump only for package amplify-category-interactions
+
+
+
+
+
 ## [2.1.11](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-interactions@2.1.10...amplify-category-interactions@2.1.11) (2020-02-13)
 
 **Note:** Version bump only for package amplify-category-interactions